requirements.txt
aioredis==1.3.1
amqp==2.6.1
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.5.0.5
boto3==1.11.5
botocore==1.14.5
celery==4.2.1
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne~=2.3
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-otp==1.0.2
django-polymorphic==3.0.0
django-redis==4.12.1
django-rest-friendship==0.1.3
django-rest-polymorphic==0.1.9
django-storages==1.10.1
djangorestframework==3.11.1
drf-yasg==1.17.1
fcm-django==0.3.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
MarkupSafe==1.1.1
msgpack==1.0.0
msgpack-python==0.5.6
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.7
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
style==1.1.0
Twisted==20.3.0
txaio==20.4.1
update==0.0.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
zope.interface==5.2.0
jsonpickle==0.7.1
cachecontrol==0.11.7
